updated the code added new functions
This is an update of the original RanchRPlauncher.bat I realized that
it would be better to just launch the programs from the Desktop links since most people
will have a link to each program there. I also added a few more functions such as check if the programs have a link
on the Desktop.